Refactor: AuthService interface 추가
